requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
yt-dlp>=2023.0.0
python-dateutil>=2.8.2
requests>=2.31.0
pyyaml>=6.0
